+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.66 (2022-11-29)
+
+**Note:** Version bump only for package @parca/components
+
 ## 0.16.65 (2022-11-22)
 
 **Note:** Version bump only for package @parca/components